src/dh/requirements.txt
pycryptodome==3.10.1
aioconsole==0.3.1
